Implemented admin role for /user/list to get full list of all users. Now all of your /person /number experience is fully personal, so no one can see your contact list

+++ demo/src/main/java/com/contacts/demo/controllers/PersonController.java
@@ -2,9 +2,12 @@ package com.contacts.demo.controllers;
 
 import com.contacts.demo.data.JpaNameRepository;
 import com.contacts.demo.data.types.Person;
+import com.contacts.demo.security.data.types.UserEntry;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.format.annotation.NumberFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -26,23 +29,24 @@ public class PersonController {
     }
 
     @GetMapping
-    public Iterable<Person> showPersons() {
-        Iterable<Person> result = nameRepositoryJPA.findAll();
+    public Iterable<Person> showPersons(@AuthenticationPrincipal UserEntry userEntry) {
+        Iterable<Person> result = nameRepositoryJPA.findAllByUid(userEntry.getUid());
         log.info("ShowPersons executed");
         return result;
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Person> showPersonById(@PathVariable("id") Integer id) {
+    public ResponseEntity<Person> showPersonById(@PathVariable("id") @NumberFormat Integer id, @AuthenticationPrincipal UserEntry userEntry) {
         Optional<Person> result;
-        if ((result = nameRepositoryJPA.findById(id)).isEmpty())
+        if ((result = nameRepositoryJPA.findByPersonIdAndUid(id, userEntry.getUid())).isEmpty())
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         log.info("ShowPersonById executed with id=" + id);
         return new ResponseEntity<>(result.get(), HttpStatus.OK);
     }
 
     @PostMapping(consumes = "application/json")
-    public Person addPerson(@RequestBody @Valid Person newPerson) {
+    public Person addPerson(@RequestBody @Valid Person newPerson, @AuthenticationPrincipal UserEntry userEntry) {
+        newPerson.setUid(userEntry.getUid());
         nameRepositoryJPA.save(newPerson);
         log.info("AddPerson executed. " + newPerson + " added");
         return newPerson;
@@ -50,12 +54,13 @@ public class PersonController {
 
     @Transactional
     @PatchMapping(path = "/{id}", consumes = "application/json")
-    public ResponseEntity<Person> editPerson(@PathVariable("id") Integer id, @RequestBody @NotEmpty Person newPerson) {
-        if (!nameRepositoryJPA.existsById(id))
+    public ResponseEntity<Person> editPerson(@PathVariable("id") @NumberFormat Integer id, @RequestBody @NotEmpty Person newPerson, @AuthenticationPrincipal UserEntry userEntry) {
+        Optional<Person> foundPerson = nameRepositoryJPA.findByPersonIdAndUid(id, userEntry.getUid());
+        if (foundPerson.isEmpty())
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
-        Person editablePerson = nameRepositoryJPA.findById(id).get();
-        String correctName = editablePerson.getName().trim();
+        Person editablePerson = foundPerson.get();
+        String correctName = editablePerson.getName();
         editablePerson.setName(correctName);
 
         if (newPerson.getName() != null)
@@ -67,8 +72,11 @@ public class PersonController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable("id") Integer id) {
+    public ResponseEntity deleteById(@PathVariable("id") @NumberFormat Integer id, @AuthenticationPrincipal UserEntry userEntry) {
+        if (!nameRepositoryJPA.existsByPersonIdAndUid(id, userEntry.getUid()))
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
         nameRepositoryJPA.deleteById(id);
         log.info("DeleteById executed. Id=" + id + " deleted");
+        return new ResponseEntity(HttpStatus.OK);
     }
 }
